pyproject.toml
[project]
name = "pipecat-dictation"
version = "0.1.0"
description = "Voice dictation and command runner"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv",
    "fastapi[all]",
    "uvicorn",
    "opencv-python",
    "pipecat-ai[openai,webrtc,silero,local,google,speechmatics,cartesia,deepgram,daily]==0.0.82",
    "pipecat-ai-small-webrtc-prebuilt",
    "aiortc",
    "nltk",
    "pyautogui",
    "python-xlib",
    "pynput",
    "pyperclip",
    "textual",
]

[tool.uv.sources]
webrtc-audio-processing = { path = "dist/webrtc_audio_processing-0.1.3.tar.gz" }
